centralize post-delete cache cleanup in a single hook
Pulls the duplicated query-cache stripping and challenge/quest
completion invalidation out of home, profile, challenge, and post
pages into usePostDeleteCleanup, called from Post itself.
