+++ src/app/(tabs)/index.tsx
@@ -9,7 +9,6 @@ import {
   Text,
   StyleSheet,
 } from "react-native";
-import { useQueryClient } from "@tanstack/react-query";
 import Post from "../../components/Post";
 import WelcomePostGroup from "../../components/WelcomePostGroup";
 import FeedSortToggle from "../../components/FeedSortToggle";
@@ -64,7 +63,6 @@ const prepareFeedItems = (posts: PostType[]): FeedItem[] => {
 const Home = () => {
   const { t } = useLanguage();
   const router = useRouter();
-  const queryClient = useQueryClient();
   const { firstTime } = useLocalSearchParams<{ firstTime?: string }>();
   const defaultSort: FeedSort = firstTime === "true" ? "popular" : "recent";
   const [, setPostCounts] = useState<Record<number, { likes: number; comments: number }>>({});
@@ -107,30 +105,6 @@ const Home = () => {
     }
   }, [refetchPosts]);
 
-  const handlePostDeleted = useCallback((post: PostType) => {
-    type HomePostsPage = { posts: PostType[]; hasMore: boolean };
-    type HomePostsData = { pages: HomePostsPage[]; pageParams: number[] };
-
-    queryClient.setQueriesData<HomePostsData>({ queryKey: ["home-posts"] }, (oldData) => {
-      if (!oldData) return oldData;
-      return {
-        ...oldData,
-        pages: oldData.pages.map((page) => ({
-          ...page,
-          posts: page.posts.filter((pagePost) => pagePost.id !== post.id),
-        })),
-      };
-    });
-
-    if (post.challenge_id) {
-      queryClient.invalidateQueries({ queryKey: ["challenge-completion", post.challenge_id] });
-    }
-
-    if (post.quest_id) {
-      queryClient.invalidateQueries({ queryKey: ["quest-completion", post.quest_id] });
-    }
-  }, [queryClient]);
-
   const handlePostCreated = useCallback(() => {
     refetchPosts();
   }, [refetchPosts]);
@@ -170,11 +144,10 @@ const Home = () => {
           post={item.post}
           postUser={postUser}
           setPostCounts={setPostCounts}
-          onPostDeleted={handlePostDeleted}
         />
       );
     },
-    [expandedWelcomeGroups, handlePostDeleted, toggleWelcomeGroup, userMap]
+    [expandedWelcomeGroups, toggleWelcomeGroup, userMap]
   );
 
   const renderEmpty = useCallback(() => {
